refactor: 배럴 파일 삭제 및 inmport 수정 (#323)
