+++ src/pages/class/components/ClassButtonWrapper/ClassButtonWrapper.tsx
@@ -3,7 +3,8 @@ import { useNavigate, useParams } from 'react-router-dom';
 import { buttonWrapperStyle } from '@/pages/class/components/ClassButtonWrapper/classButtonWrapper.css';
 import { LessonDetailResponse } from '@/pages/class/types/api';
 import { ROUTES_CONFIG } from '@/routes/routesConfig';
-import { IcHeartFilledGray07, IcHeartOutlinedGray07 } from '@/shared/assets/svg';
+import IcHeartFilledGray07 from '@/shared/assets/svg/IcHeartFilledGray07';
+import IcHeartOutlinedGray07 from '@/shared/assets/svg/IcHeartOutlinedGray07';
 import BoxButton from '@/shared/components/BoxButton/BoxButton';
 import Flex from '@/shared/components/Flex/Flex';
 import { BUTTON_TEXT, DISABLED_STATUS } from '@/shared/constants';
