+++ src/pages/class/components/ClassHeader/ClassHeader.tsx
@@ -5,7 +5,8 @@ import {
   headerTextStyle,
 } from '@/pages/class/components/ClassHeader/classHeader.css';
 import { containerStyle } from '@/pages/home/components/HomeHeader/homeHeader.css';
-import { IcBack, IcBackWhite24 } from '@/shared/assets/svg';
+import IcBack from '@/shared/assets/svg/IcBack';
+import IcBackWhite24 from '@/shared/assets/svg/IcBackWhite24';
 import Flex from '@/shared/components/Flex/Flex';
 import Head from '@/shared/components/Head/Head';
