+++ src/pages/class/components/TabWrapper/TabLocation/TabLocation.tsx
@@ -5,7 +5,7 @@ import {
   streetAddressStyle,
 } from '@/pages/class/components/TabWrapper/TabLocation/tabLocation.css';
 import { LessonDetailResponse } from '@/pages/class/types/api';
-import { IcLocation60 } from '@/shared/assets/svg';
+import IcLocation60 from '@/shared/assets/svg/IcLocation60';
 import Flex from '@/shared/components/Flex/Flex';
 import Head from '@/shared/components/Head/Head';
 import Text from '@/shared/components/Text/Text';
