+++ src/pages/home/components/BottomSection/BottomSection.tsx
@@ -3,7 +3,8 @@ import { useNavigate } from 'react-router-dom';
 import { usePostLogout } from '@/pages/auth/apis/queries';
 import * as styles from '@/pages/home/components/BottomSection/bottomSection.css';
 import { LIST_DATA } from '@/pages/mypage/constants/myPageList';
-import { IcArrowRightSmallGray0432, IcArrowRightSmallGray0732 } from '@/shared/assets/svg';
+import IcArrowRightSmallGray0432 from '@/shared/assets/svg/IcArrowRightSmallGray0432';
+import IcArrowRightSmallGray0732 from '@/shared/assets/svg/IcArrowRightSmallGray0732';
 import Divider from '@/shared/components/Divider/Divider';
 import Flex from '@/shared/components/Flex/Flex';
 import { notify } from '@/shared/components/Toast/Toast';
