+++ src/pages/instructor/classRegister/components/ClassRepresentImage/ClassRepresentImage.tsx
@@ -4,7 +4,8 @@ import {
   previewImgStyle,
 } from '@/pages/instructor/classRegister/components/ClassRepresentImage/classRepresentImage.css';
 import Description from '@/pages/instructor/classRegister/components/Description';
-import { IcPlusGray0524, IcXCircleMain0324 } from '@/shared/assets/svg';
+import IcPlusGray0524 from '@/shared/assets/svg/IcPlusGray0524';
+import IcXCircleMain0324 from '@/shared/assets/svg/IcXCircleMain0324';
 import Flex from '@/shared/components/Flex/Flex';
 import Text from '@/shared/components/Text/Text';
