+++ src/pages/onboarding/components/ProfileStep/ProfileStep.tsx
@@ -2,7 +2,7 @@ import { icCameraStyle, inputStyle, previewImgStyle } from '@/pages/onboarding/c
 import { INFO_KEY, MAX_NICKNAME_LENGTH } from '@/pages/onboarding/constants';
 import { onboardInfoTypes } from '@/pages/onboarding/types/onboardInfoTypes';
 import defaultProfile from '@/shared/assets/images/image_profile_basic.png';
-import { IcCameraMain0624 } from '@/shared/assets/svg';
+import IcCameraMain0624 from '@/shared/assets/svg/IcCameraMain0624';
 import Flex from '@/shared/components/Flex/Flex';
 import Head from '@/shared/components/Head/Head';
 import Input from '@/shared/components/Input/Input';
