+++ src/pages/search/components/TabContainer/TagSection/TagSection.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import BottomSheet from '@/pages/search/components/TabContainer/TagSection/BottomSheet/BottomSheet';
 import { tagCustomStyle } from '@/pages/search/components/TabContainer/TagSection/tagSection.css';
-import { IcFilterGray } from '@/shared/assets/svg';
+import IcFilterGray from '@/shared/assets/svg/IcFilterGray';
 import Flex from '@/shared/components/Flex/Flex';
 import Tag from '@/shared/components/Tag/Tag';
