--- src/shared/assets/svg/index.ts
@@ -1,63 +1,0 @@
-export { default as BtnCheck } from './BtnCheck'
-export { default as IcArrowDownGray1032 } from './IcArrowDownGray1032'
-export { default as IcArrowRightGray0614 } from './IcArrowRightGray0614'
-export { default as IcArrowRightSmallGray0432 } from './IcArrowRightSmallGray0432'
-export { default as IcArrowRightSmallGray0732 } from './IcArrowRightSmallGray0732'
-export { default as IcArrowRightSmallMain0332New } from './IcArrowRightSmallMain0332New'
-export { default as IcArrowRightWhite14 } from './IcArrowRightWhite14'
-export { default as IcArrowUnderGray } from './IcArrowUnderGray'
-export { default as IcArrowUpGray1032 } from './IcArrowUpGray1032'
-export { default as IcAsset1 } from './IcAsset1'
-export { default as IcAssetTop1 } from './IcAssetTop1'
-export { default as IcAssetTop2 } from './IcAssetTop2'
-export { default as IcAssetTop3 } from './IcAssetTop3'
-export { default as IcBack } from './IcBack'
-export { default as IcBackWhite24 } from './IcBackWhite24'
-export { default as IcBtnEtc } from './IcBtnEtc'
-export { default as IcCalendarcheckColor3D24 } from './IcCalendarcheckColor3D24'
-export { default as IcCalendarcheckMono3D24 } from './IcCalendarcheckMono3D24'
-export { default as IcCameraMain0624 } from './IcCameraMain0624'
-export { default as IcCheckcircleGray0524 } from './IcCheckcircleGray0524'
-export { default as IcCheckcircleMain0324 } from './IcCheckcircleMain0324'
-export { default as IcCheckGray0724 } from './IcCheckGray0724'
-export { default as IcCheckMain0324 } from './IcCheckMain0324'
-export { default as IcCheckMain0424 } from './IcCheckMain0424'
-export { default as IcClassCanceledGray0824 } from './IcClassCanceledGray0824'
-export { default as IcClassEndMain0324 } from './IcClassEndMain0324'
-export { default as IcClassIngMain0324 } from './IcClassIngMain0324'
-export { default as IcClassSoonMain0324 } from './IcClassSoonMain0324'
-export { default as IcClose } from './IcClose'
-export { default as IcCloseBlack24 } from './IcCloseBlack24'
-export { default as IcCloseWhite24 } from './IcCloseWhite24'
-export { default as IcDocumentBlack20 } from './IcDocumentBlack20'
-export { default as IcError } from './IcError'
-export { default as IcFilterGray } from './IcFilterGray'
-export { default as IcGraduationBlack20 } from './IcGraduationBlack20'
-export { default as IcHeartFilledGray07 } from './IcHeartFilledGray07'
-export { default as IcHeartOutlinedGray07 } from './IcHeartOutlinedGray07'
-export { default as IcImageTemp } from './IcImageTemp'
-export { default as IcInstagram20 } from './IcInstagram20'
-export { default as IcKakaoKakaobrown28 } from './IcKakaoKakaobrown28'
-export { default as IcLevelAdvanced } from './IcLevelAdvanced'
-export { default as IcLevelBasic } from './IcLevelBasic'
-export { default as IcLevelIntermediate } from './IcLevelIntermediate'
-export { default as IcLevelStarter } from './IcLevelStarter'
-export { default as IcLocation60 } from './IcLocation60'
-export { default as IcLogoSmallBlack } from './IcLogoSmallBlack'
-export { default as IcLogoSmallWhite } from './IcLogoSmallWhite'
-export { default as IcMypageBlack24 } from './IcMypageBlack24'
-export { default as IcMypageWhite24 } from './IcMypageWhite24'
-export { default as IcPlusGray0524 } from './IcPlusGray0524'
-export { default as IcProfileBasic } from './IcProfileBasic'
-export { default as IcQuesitonmark } from './IcQuesitonmark'
-export { default as IcSearchBlack24 } from './IcSearchBlack24'
-export { default as IcSearchGray } from './IcSearchGray'
-export { default as IcSearchWhite24 } from './IcSearchWhite24'
-export { default as IcSparkleMain20 } from './IcSparkleMain20'
-export { default as IcSpeaker3D } from './IcSpeaker3D'
-export { default as IcThunderMain0424 } from './IcThunderMain0424'
-export { default as IcXCircleGray } from './IcXCircleGray'
-export { default as IcXCircleGray0424 } from './IcXCircleGray0424'
-export { default as IcXCircleMain0324 } from './IcXCircleMain0324'
-export { default as IcXMain04 } from './IcXMain04'
-export { default as IcYoutube20 } from './IcYoutube20'
