+++ src/stories/Tag.stories.tsx
@@ -1,5 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { IcCalendarcheckColor3D24, IcCalendarcheckMono3D24 } from '@/shared/assets/svg';
+import IcCalendarcheckColor3D24 from '@/shared/assets/svg/IcCalendarcheckColor3D24';
+import IcCalendarcheckMono3D24 from '@/shared/assets/svg/IcCalendarcheckMono3D24';
 import Tag from '@/shared/components/Tag/Tag';
 
 const meta = {
